Joined date, time and color pickers to pickers.py
